Move all benchmarking within deepsparse/benchmark/
