+++ src/deepsparse/benchmark/stream_benchmark.py
